requirements.txt
PyYAML==6.0.2
requests==2.32.3
language-tool-python==2.7.1
Pillow==10.4.0
markdown==3.6
rich==13.7.1
